Added proper mouseInput
Problem with ImageLoader taking too much time to load images or returns
null.When you start the game wherever you click you get in game mode.
Sometimes imageLoader returns null and when you click on the main screen
it does not draw on game canvas but it is in gameState. Summary we have
problem with imageLoader.